Move Binaryen specific code from eei.h to binaryen.h
