+++ src/eei.cpp
@@ -14,8 +14,11 @@
  * limitations under the License.
  */
 
-#include <stdexcept>
-#include <array>
+#include <vector>
+#include <sstream>
+#include <iostream>
+#include <iomanip>
+
 #include "debugging.h"
 #include "eei.h"
 #include "exceptions.h"
